+++ lib/lightning_web/live/channel_request_live/show.ex
@@ -57,8 +57,7 @@ defmodule LightningWeb.ChannelRequestLive.Show do
             <LayoutComponents.breadcrumbs>
               <LayoutComponents.breadcrumb_project_picker
                 project={@project}
-                current_user={@current_user}
-                access_root={@access_root}
+                label={@project_label}
               />
               <LayoutComponents.breadcrumb_items items={[
                 {"History", ~p"/projects/#{@project}/history"},