+++ lib/lightning_web/live/run_live/show.ex
@@ -66,8 +66,7 @@ defmodule LightningWeb.RunLive.Show do
             <LayoutComponents.breadcrumbs>
               <LayoutComponents.breadcrumb_project_picker
                 project={@project}
-                current_user={@current_user}
-                access_root={@access_root}
+                label={@project_label}
               />
               <LayoutComponents.breadcrumb_items items={[
                 {"History", ~p"/projects/#{@project}/history"}