+++ lib/lightning_web/live/workflow_live/index.ex
@@ -41,8 +41,7 @@ defmodule LightningWeb.WorkflowLive.Index do
             <LayoutComponents.breadcrumbs>
               <LayoutComponents.breadcrumb_project_picker
                 project={@project}
-                current_user={@current_user}
-                access_root={@access_root}
+                label={@project_label}
               />
               <LayoutComponents.breadcrumb>
                 <:label>{@page_title}</:label>